[SALES_TAXES] TaxesServiceTest
Refactored to use BigDecimal instead of Double

+++ services/src/main/java/com/itemis/salestaxes/services/dto/ReceiptDTO.java
@@ -3,13 +3,15 @@ package com.itemis.salestaxes.services.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ReceiptDTO {
 
     private ShoppingBasketDTO shoppingBasket;
-    private Double salesTaxes;
-    private Double total;
+    private BigDecimal salesTaxes;
+    private BigDecimal total;
 }
