+++ services/src/main/java/com/itemis/salestaxes/services/dto/ShoppingBasketItemDTO.java
@@ -4,12 +4,14 @@ import com.itemis.salestaxes.services.domain.SaleType;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ShoppingBasketItemDTO {
     private ProductDTO product;
-    private Double price;
-    private Integer quantity;
+    private BigDecimal price;
+    private BigDecimal quantity;
     private SaleType saleType;
